requirements.txt
openai==1.25.0
tiktoken==0.7.0
Flask==2.0.2
Werkzeug==2.0.2
evtx==0.8.4
mft==0.6.1
pypdf==4.3.1
pefile==2023.2.7
python-registry==1.3.1
regipy==3.0.2
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
sentence-transformers==3.0.1
transformers==4.44.0
chromadb==0.4.24
ciso8601==2.3.1